+++ src/js/components/DetailViewHeader.js
