+++ src/js/plugin-bridge/PluginModules.js
@@ -58,6 +58,7 @@ module.exports = {
     ClusterHeader: 'ClusterHeader',
     ClusterName: 'ClusterName',
     CollapsingString: 'CollapsingString',
+    DetailViewHeader: 'DetailViewHeader',
     DCOSLogo: 'DCOSLogo',
     DescriptionList: 'DescriptionList',
     FilterBar: 'FilterBar',
@@ -70,7 +71,6 @@ module.exports = {
     MesosphereLogo: 'icons/MesosphereLogo',
     NestedServiceLinks: 'NestedServiceLinks',
     Page: 'Page',
-    PageHeader: 'PageHeader',
     RequestErrorMsg: 'RequestErrorMsg',
     StatusBar: 'StatusBar',
     TabForm: 'TabForm',
